move/rename entries
