+++ api/sync15/blobdoc.go
@@ -1,0 +1,212 @@
+package sync15
+
+import (
+	"bufio"
+	"bytes"
+	"crypto/sha256"
+	"encoding/hex"
+	"encoding/json"
+	"errors"
+	"io"
+	"io/ioutil"
+	"sort"
+	"strconv"
+	"strings"
+
+	"github.com/juruen/rmapi/archive"
+	"github.com/juruen/rmapi/log"
+	"github.com/juruen/rmapi/model"
+)
+
+type BlobDoc struct {
+	Files []*Entry
+	Entry
+	archive.MetadataFile
+}
+
+func NewBlobDoc(name, documentId, colType string) *BlobDoc {
+	return &BlobDoc{
+		MetadataFile: archive.MetadataFile{
+			DocName:        name,
+			CollectionType: colType,
+		},
+		Entry: Entry{
+			DocumentID: documentId,
+		},
+	}
+
+}
+
+func (d *BlobDoc) Rehash() error {
+
+	hash, err := HashEntries(d.Files)
+	if err != nil {
+		return err
+	}
+	log.Info.Println("New doc hash: ", hash)
+	d.Hash = hash
+	return nil
+}
+
+func (d *BlobDoc) UpdateMetadata() (hash string, reader io.Reader, err error) {
+	jsn, err := json.Marshal(d.MetadataFile)
+	if err != nil {
+		return
+	}
+	sha := sha256.New()
+	sha.Write(jsn)
+	hash = hex.EncodeToString(sha.Sum(nil))
+	log.Info.Println("new hash", hash)
+	reader = bytes.NewReader(jsn)
+	found := false
+	for _, f := range d.Files {
+		if strings.HasSuffix(f.DocumentID, ".metadata") {
+			f.Hash = hash
+			found = true
+			break
+		}
+	}
+	if !found {
+		err = errors.New("metadata not found")
+	}
+
+	return
+}
+
+func (d *BlobDoc) AddFile(e *Entry) error {
+	d.Files = append(d.Files, e)
+	return d.Rehash()
+}
+
+func (t *Tree) Add(d *BlobDoc) error {
+	if len(d.Files) == 0 {
+		return errors.New("no files")
+	}
+	t.Docs = append(t.Docs, d)
+	return t.Rehash()
+}
+
+func (t *BlobDoc) IndexReader() (io.ReadCloser, error) {
+	if len(t.Files) == 0 {
+		return nil, errors.New("no files")
+	}
+	pipeReader, pipeWriter := io.Pipe()
+	w := bufio.NewWriter(pipeWriter)
+	go func() {
+		defer pipeWriter.Close()
+		w.WriteString(SchemaVersion)
+		w.WriteString("\n")
+		for _, d := range t.Files {
+			w.WriteString(d.Line())
+			w.WriteString("\n")
+		}
+		w.Flush()
+	}()
+
+	return pipeReader, nil
+}
+
+// Extract the documentname from metadata blob
+func (doc *BlobDoc) SyncName(fileEntry *Entry, r RemoteStorage) error {
+	if strings.HasSuffix(fileEntry.DocumentID, ".metadata") {
+		log.Info.Println("Reading metadata: " + doc.DocumentID)
+
+		metadata := archive.MetadataFile{}
+
+		meta, err := r.GetReader(fileEntry.Hash)
+		if err != nil {
+			return err
+		}
+		defer meta.Close()
+		content, err := ioutil.ReadAll(meta)
+		if err != nil {
+			return err
+		}
+		err = json.Unmarshal(content, &metadata)
+		if err != nil {
+			log.Info.Printf("cannot read metadata %v", err)
+		}
+		log.Info.Println(metadata.DocName)
+		doc.MetadataFile = metadata
+	}
+
+	return nil
+}
+
+func (d *BlobDoc) Line() string {
+	var sb strings.Builder
+	if d.Hash == "" {
+		log.Error.Print("missing hash for: ", d.DocumentID)
+	}
+	sb.WriteString(d.Hash)
+	sb.WriteRune(Delimiter)
+	sb.WriteString(DocType)
+	sb.WriteRune(Delimiter)
+	sb.WriteString(d.DocumentID)
+	sb.WriteRune(Delimiter)
+
+	numFilesStr := strconv.Itoa(len(d.Files))
+	sb.WriteString(numFilesStr)
+	sb.WriteRune(Delimiter)
+	sb.WriteString("0")
+	return sb.String()
+}
+
+func (doc *BlobDoc) Sync(e *Entry, r RemoteStorage) error {
+	doc.Entry = *e
+	entryIndex, err := r.GetReader(e.Hash)
+	if err != nil {
+		return err
+	}
+	defer entryIndex.Close()
+	entries, err := parseIndex(entryIndex)
+	if err != nil {
+		return err
+	}
+
+	head := make([]*Entry, 0)
+	current := make(map[string]*Entry)
+	new := make(map[string]*Entry)
+
+	for _, e := range entries {
+		new[e.DocumentID] = e
+	}
+
+	for _, currentEntry := range doc.Files {
+		if newEntry, ok := new[currentEntry.DocumentID]; ok {
+			if newEntry.Hash != currentEntry.Hash {
+				err = doc.SyncName(newEntry, r)
+				if err != nil {
+					return err
+				}
+				currentEntry.Hash = newEntry.Hash
+			}
+			head = append(head, currentEntry)
+			current[currentEntry.DocumentID] = currentEntry
+		}
+	}
+
+	for k, newEntry := range new {
+		if _, ok := current[k]; !ok {
+			err = doc.SyncName(newEntry, r)
+			if err != nil {
+				return err
+			}
+			head = append(head, newEntry)
+		}
+	}
+	sort.Slice(head, func(i, j int) bool { return head[i].DocumentID < head[j].DocumentID })
+	doc.Files = head
+	return nil
+
+}
+func (doc *BlobDoc) ToDocument() *model.Document {
+	return &model.Document{
+		ID:           doc.DocumentID,
+		VissibleName: doc.MetadataFile.DocName,
+		Version:      doc.MetadataFile.Version,
+		Parent:       doc.MetadataFile.Parent,
+		Type:         doc.MetadataFile.CollectionType,
+	}
+
+}
